// src/features/cities/interfaces/global-stats.interface.ts
/**
 * Interfaz para las estadísticas globales de la plataforma
 */
export interface GlobalStats {
  totalCities: number
  totalUsers: number
  totalPhotos: number
  totalVideos: number
  totalMedia: number
  totalPosts: number
  totalBusinesses: number
  totalServices: number
  totalDirectoryEntries: number
}